gateway/requirements.txt
Flask
requests
python-dotenv
PyJWT
Werkzeug
Flask-Limiter
redis
Flask-RESTx
pybreaker
prometheus_client
python-consul
pytest
dependencies